Move Table out of the Vector hierarchy
